Updated to newly forked version of plugin
