+++ .gitignore
@@ -1,2 +1,3 @@
 /.gradle/
 /build/
+.idea/
