+++ settings.gradle
@@ -1,2 +1,2 @@
 rootProject.name = 'gradle-gatling-plugin'
-
+enableFeaturePreview('STABLE_PUBLISHING')
