+++ src/main/resources/META-INF/gradle-plugins/gatling.properties
@@ -1,1 +1,1 @@
-implementation-class=uk.gov.gradle.plugins.GatlingPlugin
+implementation-class=com.github.cmhdave.gradle.plugins.GatlingPlugin
